move AmountForm to containers
